+++ src/strand-menu/strand-menu.js
@@ -12,10 +12,24 @@
 			StrandTraits.Refable
 		],
 
+		listeners: {
+			'click':'_handleSelect'
+		},
+
 		attached: function() {
 			this.async(function() {
 				if (this._target) this._target.style.cursor = 'pointer';
 			});
+		},
+
+		_handleSelect: function(e) {
+
+			if (e.target !== this.$.container) {
+				this.fire('selected',{
+					item: e.target,
+					value: e.target.getAttribute('value') || Polymer.dom(e.target).textContent.trim()
+				});
+			}
 		}
 	});